Replace the download URL to get rid of cachito-testing/cachi2 reference
When this test was created, an artificial release was created in the
dedicated (back then) repository for the test only to reference it in
the lockfile for test purposes. In this particular instance the URL is
supposed to return a 404, hence use a fake one and drop cachito/cachi2
references.

Signed-off-by: Erik Skultety <eskultet@redhat.com>