Replace the download URLs in the lockfile to point to our own tarballs
When this test was created, an artificial release was created in the
dedicated (back then) repository for the test only to reference it in
the lockfile for test purposes. We consolidated all integration test
cases to a single repository with the origin potentially removed in the
future, so to make tests stable, we should point them to our own
regular release tarballs.

Signed-off-by: Erik Skultety <eskultet@redhat.com>