Drop Containerfile
This is not the actual Containerfile that is used during integration
test runs, these are still hosted in the main cachi2 repo. Drop this
one to avoid confusion.

Note that a Containerfile will be added in the future as part of a
larger Containerfile migration to the respective test branches.

Signed-off-by: Erik Skultety <eskultet@redhat.com>