Add pip/legacy-cachito-hash integration test
This test should replace pip/mixed-hashes which contains legacy cachito
hash qualifier with manually added standard sha256 hash to an URL
dependency.

The functionality is still supported in our pip backend, so it is worth
having a simple integration test. Once we remove it from Hermeto, this
test can be removed as well.

There is only one dependency in `requirements.txt` without any Rust
extensions, which happened to be in pip/mixed-hashes test.

Signed-off-by: Michal Šoltis <msoltis@redhat.com>